Implementation de l'extension sous OpenLayers. + Mutualisation du code du DOM (fonction de creation des profils) avec Leaflet. + Exemples
