+++ samples-src/templates/ol3/ol-sample-amd-layout.hbs
@@ -5,6 +5,8 @@
         {{/extend}}
         {{#extend "partials-ol-common-head"}}
         {{/extend}}
+        {{#block "vendor"}}
+        {{/block}}
         {{#extend "partials-ol-amd-head"}}
         {{/extend}}
