+++ samples-src/templates/ol3/ol-sample-bundle-layout.hbs
@@ -5,6 +5,8 @@
         {{/extend}}
         {{#extend "partials-ol-common-head"}}
         {{/extend}}
+        {{#block "vendor"}}
+        {{/block}}
         {{#extend "partials-ol-bundle-head"}}
         {{/extend}}
